Fix timeline scrolling
Fixes #72

This uses a special hook to make sure we scroll down automatically after new event tiles are rendered, assuming the user was previously at the bottom of the timeline.

Note: The scroll up event does *not* fire when we add a new event tile. It appears as though perfect-scrollbar doesn't fire any events when the content gets longer, and we rely on this behaviour.