Add an early structure for notification processing
The general idea is that the room service will call into the notification service for each renderable event. This offloads push rule calculations to the notification service.

Notifications and unread counts are both going to be calculated client side. Therefore we can ignore the server values in the sync response. This is because issues like [https://github.com/vector-im/riot-web/issues/3363] are inevitable if we try and do a mix of   server and client processing. 

However, to get to that point we need read receipts so we can see what we've read and what we haven't read. Ideally this reliance on read receipts will disappear when this is resolved: https://github.com/vector-im/riot-meta/issues/66

The amount of interaction between the room service and notification service is still relatively undecided. It will call into the notification service for determining if something should notify, however the read receipt parsing needs to somehow update the notification count somewhere.